split into components and better footer
